+++ src/main/java/org/myteam/server/board/service/BoardCommentService.java
@@ -12,6 +12,7 @@ import org.myteam.server.board.dto.request.BoardCommentUpdateRequest;
 import org.myteam.server.board.repository.BoardCommentRepository;
 import org.myteam.server.board.repository.BoardReplyRepository;
 import org.myteam.server.chat.domain.BadWordFilter;
+import org.myteam.server.global.util.upload.MediaUtils;
 import org.myteam.server.member.entity.Member;
 import org.myteam.server.member.service.MemberReadService;
 import org.myteam.server.member.service.SecurityReadService;
@@ -84,7 +85,7 @@ public class BoardCommentService {
         boardComment.verifyBoardCommentAuthor(boardComment, member);
 
         // S3 이미지 삭제
-        s3Service.deleteFile(s3Service.getImagePath(boardComment.getImageUrl()));
+        s3Service.deleteFile(MediaUtils.getImagePath(boardComment.getImageUrl()));
         // 대댓글 삭제 (카운트도 포함)
         int minusCount = deleteBoardReply(boardComment.getId());
         // 댓글 삭제
@@ -100,7 +101,7 @@ public class BoardCommentService {
     private int deleteBoardReply(Long boardCommentId) {
         List<BoardReply> boardReplyList = boardReplyReadService.findByBoardCommentId(boardCommentId);
         for (BoardReply boardReply : boardReplyList) {
-            s3Service.deleteFile(s3Service.getImagePath(boardReply.getImageUrl()));
+            s3Service.deleteFile(MediaUtils.getImagePath(boardReply.getImageUrl()));
             boardReplyRepository.delete(boardReply);
         }
         return boardReplyList.size();
@@ -111,7 +112,7 @@ public class BoardCommentService {
      */
     private void verifyBoardCommentImageAndRequestImage(String boardCommentImageUrl, String requestImageUrl) {
         if (!boardCommentImageUrl.equals(requestImageUrl)) {
-            s3Service.deleteFile(s3Service.getImagePath(requestImageUrl));
+            s3Service.deleteFile(MediaUtils.getImagePath(requestImageUrl));
         }
     }
 }
