+++ src/main/java/org/myteam/server/board/service/BoardReplyService.java
@@ -7,6 +7,7 @@ import org.myteam.server.board.dto.reponse.BoardReplyResponse;
 import org.myteam.server.board.dto.request.BoardReplySaveRequest;
 import org.myteam.server.board.repository.BoardReplyRepository;
 import org.myteam.server.chat.domain.BadWordFilter;
+import org.myteam.server.global.util.upload.MediaUtils;
 import org.myteam.server.member.entity.Member;
 import org.myteam.server.member.service.MemberReadService;
 import org.myteam.server.member.service.SecurityReadService;
@@ -80,7 +81,7 @@ public class BoardReplyService {
 
         boardReply.verifyBoardReplyAuthor(boardReply, loginUser);
 
-        s3Service.deleteFile(s3Service.getImagePath(boardReply.getImageUrl()));
+        s3Service.deleteFile(MediaUtils.getImagePath(boardReply.getImageUrl()));
         boardReplyRepository.delete(boardReply);
 
         boardCountService.minusCommentCount(boardReply.getBoardComment().getBoard().getId());
@@ -91,7 +92,7 @@ public class BoardReplyService {
      */
     private void verifyBoardReplyImageAndRequestImage(String boardReplyImageUrl, String requestImageUrl) {
         if (!boardReplyImageUrl.equals(requestImageUrl)) {
-            s3Service.deleteFile(s3Service.getImagePath(requestImageUrl));
+            s3Service.deleteFile(MediaUtils.getImagePath(requestImageUrl));
         }
     }
 }
