+++ src/main/java/org/myteam/server/upload/service/S3Service.java
@@ -117,13 +117,4 @@ public class S3Service {
         }
         return false;
     }
-
-    public static String getImagePath(String url) {
-        String target = "devbucket/";
-        int index = url.indexOf(target);
-        if (index != -1) {
-            return url.substring(index + target.length());
-        }
-        return null;
-    }
 }
